+++ engine/flink/management/periodic/src/main/scala/pl/touk/nussknacker/engine/management/periodic/DeploymentActor.scala
@@ -7,6 +7,7 @@ import pl.touk.nussknacker.engine.management.periodic.model.PeriodicProcessDeplo
 
 import scala.concurrent.Future
 import scala.concurrent.duration._
+import scala.util.{Failure, Success}
 
 object DeploymentActor {
 
@@ -24,7 +25,7 @@ object DeploymentActor {
 
   private case class WaitingForDeployment(ids: List[PeriodicProcessDeployment])
 
-  private case class DeploymentCompleted(success: Boolean)
+  private case object DeploymentCompleted
 }
 
 class DeploymentActor(findToBeDeployed: => Future[Seq[PeriodicProcessDeployment]],
@@ -42,19 +43,31 @@ class DeploymentActor(findToBeDeployed: => Future[Seq[PeriodicProcessDeployment]
 
   override def receive: Receive = {
     case CheckToBeDeployed =>
-      logger.debug("Checking scenarios to be deployed")
-      findToBeDeployed.foreach { runDetails => self ! WaitingForDeployment(runDetails.toList) }
+      logger.trace("Checking scenarios to be deployed")
+      findToBeDeployed.onComplete {
+        case Success(runDetailsSeq) =>
+          logger.debug(s"Found ${runDetailsSeq.size} to be deployed: ${runDetailsSeq.map(_.display)}")
+          self ! WaitingForDeployment(runDetailsSeq.toList)
+        case Failure(exception) =>
+          logger.error("Finding scenarios to be deployed failed unexpectedly", exception)
+      }
     case WaitingForDeployment(Nil) =>
     case WaitingForDeployment(runDetails :: _) =>
-      logger.info("Found a scenario to be deployed: {}", runDetails.id)
+      logger.info(s"Found a scenario to be deployed: ${runDetails.display}")
       context.become(receiveOngoingDeployment(runDetails))
-      deploy(runDetails) onComplete { result => self ! DeploymentCompleted(result.isSuccess) }
+      deploy(runDetails) onComplete {
+        case Success(_) =>
+          self ! DeploymentCompleted
+        case Failure(exception) =>
+          logger.error(s"Deployment of ${runDetails.display} failed unexpectedly", exception)
+          self ! DeploymentCompleted
+      }
   }
 
   private def receiveOngoingDeployment(runDetails: PeriodicProcessDeployment): Receive = {
     case CheckToBeDeployed =>
-    case DeploymentCompleted(success) =>
-      logger.info("Deployment {} completed, success: {}", runDetails.id, success)
+      logger.debug(s"Still waiting for ${runDetails.display} to be deployed")
+    case DeploymentCompleted =>
       context.unbecome()
   }
 }