+++ docs/Changelog.md
@@ -36,6 +36,8 @@
   * Handling of API Keys in query parameter and cookie
   * It's possible to configure which HTTP codes (404 by default) can be used as successful, empty response
   * Documentation link is taken from global configuration, if operation doesn't provide one
+* [#3730](https://github.com/TouK/nussknacker/pull/3730) Fix: Trimming additional fields at JsonToNuStruct when additionalProperties is enabled
+* [#3733](https://github.com/TouK/nussknacker/pull/3733) Fix for: some validation (e.g. Flink scenario name validation) were causing error message blinking in scenario properties. 
 
 1.6.1 (8 Nov 2022)
 ------------------------